+++ statemachine/transition.py
@@ -4,6 +4,7 @@ from functools import wraps
 from weakref import ref
 
 from .callbacks import Callbacks, ConditionWrapper, methodcaller, resolver_factory
+from .events import Events
 
 
 class Transition(object):
@@ -15,7 +16,7 @@ class Transition(object):
         self,
         source,
         destination,
-        trigger=None,
+        event=None,
         validators=None,
         conditions=None,
         unless=None,
@@ -25,7 +26,7 @@ class Transition(object):
     ):
         self.source = source
         self.destination = destination
-        self.trigger = trigger
+        self._events = Events().add(event)
         self.validators = Callbacks().add(validators)
         self.before = (
             Callbacks()
@@ -42,8 +43,8 @@ class Transition(object):
         self.machine = None
 
     def __repr__(self):
-        return "{}({!r}, {!r}, trigger={!r})".format(
-            type(self).__name__, self.source, self.destination, self.trigger
+        return "{}({!r}, {!r}, event={!r})".format(
+            type(self).__name__, self.source, self.destination, self.event
         )
 
     def _get_promisse_to_machine(self, func):
@@ -66,13 +67,18 @@ class Transition(object):
 
         self.before.add(
             [
-                "before_{}".format(self.trigger),
-                "on_{}".format(self.trigger),
+                pattern.format(event)
+                for pattern in ["before_{}", "on_{}"]
+                for event in self._events
             ],
             suppress_errors=True,
         )
         self.after.add(
-            ["after_{}".format(self.trigger), "after_transition"],
+            [
+                pattern.format(event)
+                for pattern in ["after_{}"]
+                for event in self._events
+            ] + ["after_transition"],
             suppress_errors=True,
         )
 
@@ -82,12 +88,26 @@ class Transition(object):
             for condition in self.conditions
         )
 
+    def match(self, event):
+        return self._events.match(event)
+
     @property
     def identifier(self):
         warnings.warn(
-            "identifier is deprecated. Use `trigger` instead", DeprecationWarning
+            "identifier is deprecated. Use `event` instead", DeprecationWarning
         )
-        return self.trigger
+        return self.event
+
+    @property
+    def event(self):
+        return str(self._events)
+
+    @property
+    def events(self):
+        return self._events
+
+    def add_event(self, value):
+        self._events.add(value)
 
     def execute(self, event_data):
         self.validators(*event_data.args, **event_data.extended_kwargs)