+++ statemachine/transition_list.py
@@ -1,3 +1,4 @@
+from collections import OrderedDict
 from .utils import ensure_iterable
 
 
@@ -33,3 +34,17 @@ class TransitionList(object):
             func = transition._get_promisse_to_machine(f)
             transition.before.add(func)
         return self
+
+    def add_event(self, event):
+        for transition in self.transitions:
+            transition.add_event(event)
+
+    @property
+    def unique_events(self):
+        # Compat Python2.7: Using OrderedDict to get a unique ordered list
+        tmp_list = OrderedDict()
+        for transition in self.transitions:
+            for event in transition.events:
+                tmp_list[event] = True
+
+        return list(tmp_list.keys())