+++ tests/examples/traffic_light_machine.py
@@ -11,7 +11,7 @@ class TrafficLightMachine(StateMachine):
 
     def on_cycle(self, event_data=None):
         return "Running {} from {} to {}".format(
-            event_data.event.name,
+            event_data.event,
             event_data.transition.source.identifier,
             event_data.transition.destination.identifier,
         )