+++ tests/test_transitions.py
@@ -11,13 +11,13 @@ from .models import MyModel
 
 def test_transition_representation(campaign_machine):
     s = repr(
-        [t for t in campaign_machine.draft.transitions if t.trigger == "produce"][0]
+        [t for t in campaign_machine.draft.transitions if t.event == "produce"][0]
     )
     assert s == (
         "Transition("
         "State('Draft', identifier='draft', value='draft', initial=True, final=False), "
         "State('Being produced', identifier='producing', value='producing', "
-        "initial=False, final=False), trigger='produce')"
+        "initial=False, final=False), event='produce')"
     )
 
 
@@ -31,19 +31,6 @@ def test_list_state_transitions(classic_traffic_light_machine):
     machine = classic_traffic_light_machine()
     transitions = [t.identifier for t in machine.green.transitions]
     assert transitions == ["slowdown"]
-
-
-def test_list_transitions_validators(classic_traffic_light_machine):
-    machine = classic_traffic_light_machine()
-
-    def custom_validator(*args, **kwargs):
-        if "weapon" not in kwargs:
-            raise LookupError("Weapon not found.")
-
-    machine.slowdown.validators = [custom_validator]
-
-    validators = [validator for validator in machine.slowdown.validators]
-    assert validators == [custom_validator]
 
 
 def test_transition_should_accept_decorator_syntax(traffic_light_machine):
@@ -202,5 +189,5 @@ def test_should_transition_with_a_dict_as_return():
 
     machine = ApprovalMachine()
 
-    result = machine.run("accept")
+    result = machine.send("accept")
     assert result == expected_result